Move Stat commands to a separate file, avoid exporting File/Dir
Move the Stat commands to a separate file to distinguish them from core
definitions. Also make File and Dir private; there's no benefit to being
able to refer to them because their contents are all private and they're
created internally.
