+++ volume/file.go
@@ -10,16 +10,16 @@ import (
 // contentCB accepts a path and returns the content associated with that path.
 type contentCB = func(context.Context, string) (plugin.SizedReader, error)
 
-// File represents a file in a volume that has content we can access.
-type File struct {
+// file represents a file in a volume that has content we can access.
+type file struct {
 	plugin.EntryBase
 	contentcb contentCB
 	path      string
 }
 
 // newFile creates a VolumeFile.
-func newFile(name string, attr plugin.EntryAttributes, cb contentCB, path string) *File {
-	vf := &File{
+func newFile(name string, attr plugin.EntryAttributes, cb contentCB, path string) *file {
+	vf := &file{
 		EntryBase: plugin.NewEntry(name),
 		contentcb: cb,
 		path:      path,
@@ -31,6 +31,6 @@ func newFile(name string, attr plugin.EntryAttributes, cb contentCB, path string
 }
 
 // Open returns the content of the file as a SizedReader.
-func (v *File) Open(ctx context.Context) (plugin.SizedReader, error) {
+func (v *file) Open(ctx context.Context) (plugin.SizedReader, error) {
 	return v.contentcb(ctx, v.path)
 }
